created custom component for shared wallet form and wallet create steps
