+++ src/components/organization/WalletSpinup.tsx
@@ -11,7 +11,7 @@ import {
 	spinupDedicatedAgent,
 	spinupSharedAgent,
 } from '../../api/organization';
-import React, { ReactElement, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 
 import type { AxiosResponse } from 'axios';
 import DedicatedIllustrate from './DedicatedIllustrate';
@@ -20,18 +20,12 @@ import SOCKET from '../../config/SocketConfig';
 import SharedIllustrate from './SharedIllustrate';
 import { nanoid } from 'nanoid';
 import {
-	createPolygonKeyValuePair,
-	getLedgerConfig,
 	getLedgers,
 } from '../../api/Agent';
 import { AlertComponent } from '../AlertComponent';
-import CopyDid from '../../commonComponents/CopyDid';
 import { DidMethod } from '../../common/enums';
-import GenerateBtnPolygon from './walletCommonComponents/GenerateBtnPolygon';
-import SetPrivateKeyValue from './walletCommonComponents/SetPrivateKeyValue';
-import SetDomainValueInput from './walletCommonComponents/SetDomainValueInput';
-import TokenWarningMessage from './walletCommonComponents/TokenWarningMessage';
-import LedgerLessMethodsComponents from './walletCommonComponents/LegderLessMethods'
+import WalletSteps from './WalletSteps';
+import SharedAgentForm from './SharedAgentForm';
 
 
 interface Values {
@@ -132,593 +126,6 @@ const NetworkInput = ({ formikHandlers }) => {
 				<span className="text-red-500 text-xs">
 					{formikHandlers?.errors?.network}
 				</span>
-			)}
-		</div>
-	);
-};
-
-const SharedAgentForm = ({
-	orgName,
-	seeds,
-	loading,
-	submitSharedWallet,
-}: ISharedAgentForm) => {
-	const [haveDidShared, setHaveDidShared] = useState(false);
-	const [networks, setNetworks] = useState([]);
-	const [selectedLedger, setSelectedLedger] = useState('');
-	const [seedVal, setSeedVal] = useState('');
-	const [selectedNetwork, setSelectedNetwork] = useState('');
-	const [selectedDid, setSelectedDid] = useState('');
-	const [mappedData, setMappedData] = useState(null);
-	const [generatedKeys, setGeneratedKeys] = useState<IPolygonKeys | null>(null);
-	const [domainValue, setDomainValue] = useState<string>('');
-	const [endPointValue, setEndPointValue] = useState<string>('');
-	const [privateKeyValue, setPrivateKeyValue] = useState<string>('');
-
-	const fetchLedgerConfig = async () => {
-		try {
-			const { data } = await getLedgerConfig();
-			if (data?.statusCode === apiStatusCodes.API_STATUS_SUCCESS) {
-				const obj = {};
-				data.data.forEach((item) => {
-					obj[item.name.toLowerCase()] = { ...item.details };
-				});
-				setMappedData(obj);
-			}
-		} catch (err) {
-			console.log('Fetch Network ERROR::::', err);
-		}
-	};
-	
-
-	const handleLedgerChange = (e) => {
-		setSelectedLedger(e.target.value);
-		setSelectedNetwork('');
-		setSelectedDid('');
-	};
-
-	const generatePolygonKeyValuePair = async () => {
-		try {
-			const orgId = await getFromLocalStorage(storageKeys.ORG_ID);
-			const resCreatePolygonKeys = await createPolygonKeyValuePair(orgId);
-			const { data } = resCreatePolygonKeys as AxiosResponse;
-
-			if (data?.statusCode === apiStatusCodes.API_STATUS_CREATED) {
-				setGeneratedKeys(data?.data);
-			}
-		} catch (err) {
-			console.log('Generate private key ERROR::::', err);
-		}
-	};
-
-	const getLedgerList = async () => {
-		const res = await fetchNetworks();
-		setNetworks(res);
-	};
-	useEffect(() => {
-		getLedgerList();
-		fetchLedgerConfig();
-	}, []);
-
-	useEffect(() => {
-		setSeedVal(seeds)
-	}, [seeds])
-
-	const showMethod = (method: string): ReactElement => {
-		switch (method) {
-			case DidMethod.POLYGON: {
-				return mappedData && selectedLedger && selectedDid ? (
-					<span>{mappedData[selectedLedger][selectedDid] || ''}</span>
-				) : (
-					<></>
-				);
-			}
-			case DidMethod.INDY: {
-				return mappedData &&
-					selectedLedger &&
-					selectedNetwork &&
-					selectedDid ? (
-					<span>
-						{mappedData[selectedLedger][selectedNetwork][selectedDid] || ''}
-					</span>
-				) : (
-					<></>
-				);
-			}
-
-			case DidMethod.KEY:
-			case DidMethod.WEB: {
-				return mappedData && selectedLedger ? (
-					<span>{mappedData[selectedLedger][selectedLedger] || ''}</span>
-				) : (
-					<></>
-				);
-			}
-			default:
-				return <></>;
-		}
-	};
-
-	const validations = {
-		label: yup.string().required('Wallet label is required'),
-		method: yup.string().required('Method is required'),
-		...(DidMethod.INDY === selectedLedger || DidMethod.POLYGON === selectedLedger) && { network: yup.string().required('Network is required') },
-		...(DidMethod.INDY === selectedLedger) && { ledger: yup.string().required('Ledger is required') },
-	}
-
-	return (
-		<div className="mt-4 max-w-lg flex-col gap-4">
-			<div className="flex items-center gap-2 mt-4">
-				<Checkbox
-					id="haveDidShared"
-					onChange={(e) => setHaveDidShared(e.target.checked)}
-				/>
-				<Label className="flex" htmlFor="haveDidShared">
-					<p>Already have DID?</p>
-				</Label>
-			</div>
-			{!haveDidShared ? (
-				<>
-					<div className="my-3">
-						<div className="block text-sm font-medium text-gray-900 dark:text-white mb-1">
-							<Label value="Seed" />
-						</div>
-						<div className="flex">
-							<CopyDid
-								className="align-center block text-sm text-gray-900 dark:text-white"
-								value={seedVal}
-							/>
-						</div>
-					</div>
-
-					<Formik
-						initialValues={{
-							seed: seedVal,
-							method: '',
-							network: '',
-							did: '',
-							ledger: '',
-							label: orgName || '',
-						}}
-						validationSchema={yup.object().shape(validations)}
-						onSubmit={(values: ValuesShared) => {
-							submitSharedWallet(
-								values,
-								privateKeyValue,
-								domainValue,
-								endPointValue,
-							);
-							
-						}}
-					>
-						{(formikHandlers) => (
-							<Form className="">
-								<div className="mb-3 relative">
-									<label
-										htmlFor="method"
-										className="text-sm font-medium text-gray-900 dark:text-gray-300"
-									>
-										Method
-										<span className="text-red-500 text-xs">*</span>
-									</label>
-									<select
-										onChange={(e) => {
-											formikHandlers.handleChange(e);
-											handleLedgerChange(e);
-											setSeedVal(seeds)
-											setSelectedNetwork('');
-											setSelectedDid('');
-											setGeneratedKeys(null);
-										}}
-										id="method"
-										name="method"
-										className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500 h-11"
-									>
-										<option value="">Select Method</option>
-										{mappedData &&
-											Object.keys(mappedData)?.map((ledger) => (
-												<option key={ledger} value={ledger}>
-													{ledger}
-												</option>
-											))}
-									</select>
-									{formikHandlers?.errors?.method &&
-										formikHandlers?.touched?.method && (
-											<span className="absolute botton-0 text-red-500 text-xs">
-												{formikHandlers?.errors?.method}
-											</span>
-										)}
-								</div>
-
-								{formikHandlers.values.method === DidMethod.POLYGON && (
-									
-									<GenerateBtnPolygon  generatePolygonKeyValuePair={()=>generatePolygonKeyValuePair()}/>
-								)}
-
-								{generatedKeys && (
-									<div className="my-3 relative">
-										<p className="text-sm truncate">
-											<span className="font-semibold text-gray-900 dark:text-white">
-												Private Key:
-											</span>
-											<div className="flex ">
-												<CopyDid
-													className="align-center block text-sm text-gray-900 dark:text-white truncate"
-													value={generatedKeys?.privateKey}
-												/>
-											</div>
-										</p>
-
-										<p className="text-sm truncate">
-											<span className="font-semibold text-gray-900 dark:text-white">
-												Address:
-											</span>
-											<div className="flex ">
-												<CopyDid
-													className="align-center block text-sm text-gray-900 dark:text-white truncate"
-													value={generatedKeys?.address}
-												/>
-											</div>
-										</p>
-									</div>
-								)}
-
-								{generatedKeys &&
-									formikHandlers.values.method === DidMethod.POLYGON && (
-										
-										<TokenWarningMessage />
-									)}
-
-								{formikHandlers.values.method === DidMethod.POLYGON && (
-
-                   <SetPrivateKeyValue setPrivateKeyValue={(val:string)=>setPrivateKeyValue(val)} privateKeyValue={privateKeyValue}/>
-									
-								)}
-
-								{formikHandlers.values.method === DidMethod.WEB && (
-									
-
-									<SetDomainValueInput setDomainValue={(val:string)=>setDomainValue(val)} domainValue={domainValue}/>
-
-								)}
-
-								{formikHandlers.values.method !== DidMethod.POLYGON &&
-									formikHandlers.values.method !== DidMethod.KEY &&
-									formikHandlers.values.method !== DidMethod.WEB && (
-										<div className="my-3 relative">
-											<label
-												htmlFor="ledger"
-												className="text-sm font-medium text-gray-900 dark:text-gray-300"
-											>
-												Ledger
-												<span className="text-red-500 text-xs">*</span>
-											</label>
-											<select
-												onChange={(e) => {
-													formikHandlers.handleChange(e);
-													setSelectedNetwork(e.target.value);
-													setSelectedDid('');
-												}}
-												value={selectedNetwork}
-												id="ledger"
-												name="ledger"
-												className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500 h-11"
-											>
-												<option value="">Select Ledger</option>
-												{mappedData &&
-													selectedLedger &&
-													mappedData[selectedLedger] &&
-													Object.keys(mappedData[selectedLedger])?.map(
-														(ledger) => (
-															<option key={ledger} value={ledger}>
-																{ledger}
-															</option>
-														),
-													)}
-											</select>
-											{formikHandlers?.errors?.ledger &&
-												formikHandlers?.touched?.ledger && (
-													<span className="absolute botton-0 text-red-500 text-xs">
-														{formikHandlers?.errors?.ledger}
-													</span>
-												)}
-										</div>
-									)}
-								{formikHandlers.values.method !== DidMethod.WEB &&
-									formikHandlers.values.method !== DidMethod.KEY && (
-										
-										< LedgerLessMethodsComponents formikHandlers={formikHandlers} setSelectedDid={(val: string) => setSelectedDid(val)} selectedDid={selectedDid} mappedData={mappedData} selectedLedger={selectedLedger} selectedNetwork={selectedNetwork} />
-									)}
-
-								<div className="my-3 relative">
-									<label
-										htmlFor="DID Method"
-										className="text-sm font-medium text-gray-900 dark:text-gray-300"
-									>
-										DID Method
-									</label>
-									<div className="bg-gray-100 font-semibold text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500 h-11">
-										{showMethod(formikHandlers.values.method)}
-									</div>
-								</div>
-								<div className="mt-3 relative">
-									<Label htmlFor="name" value="Wallet Label" />
-									<span className="text-red-500 text-xs">*</span>
-
-									<Field
-										id="label"
-										name="label"
-										className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
-										type="text"
-									/>
-									{formikHandlers?.errors?.label &&
-										formikHandlers?.touched?.label && (
-											<span className="text-red-500 absolute text-xs">
-												{formikHandlers?.errors?.label}
-											</span>
-										)}
-								</div>
-								<div className="w-full flex justify-end">
-									<Button
-										type="submit"
-										className="flex h-min p-0.5 focus:z-10 focus:outline-none border border-transparent enabled:hover:bg-cyan-800 dark:enabled:hover:bg-cyan-700 mt-4 text-base font-medium text-center text-white bg-primary-700 rounded-md hover:!bg-primary-800 focus:ring-4 focus:ring-primary-300 sm:w-auto dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800"
-									>
-										Submit
-									</Button>
-								</div>
-							</Form>
-						)}
-					</Formik>
-				</>
-			) : (
-				<Formik
-					initialValues={{
-						seed: '',
-						method: '',
-						network: '',
-						did: '',
-						ledger: '',
-						label: orgName || '',
-					}}
-					validationSchema={yup.object().shape({
-						seed: yup.string().required('Seed is required'),
-						network: yup.string().required('Network is required'),
-						label: yup.string().required('Wallet label is required'),
-					})}
-					onSubmit={(values: ValuesShared) => {
-						submitSharedWallet(
-							values,
-							privateKeyValue,
-							domainValue,
-							endPointValue,
-						);
-					}}
-				>
-					{(formikHandlers) => (
-						<Form className="">
-							<div className="mt-3 relative">
-								<div className="block">
-									<Label htmlFor="seed" value="Seed" />
-									<span className="text-red-500 text-xs">*</span>
-								</div>
-
-								<Field
-									id="seed"
-									name="seed"
-									className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
-									type="text"
-								/>
-								{formikHandlers?.errors?.seed &&
-									formikHandlers?.touched?.seed && (
-										<span className="text-red-500 absolute text-xs">
-											{formikHandlers?.errors?.seed}
-										</span>
-									)}
-							</div>
-
-							<div className="mt-3 relative">
-								<div className="block">
-									<Label htmlFor="did" value="DID" />
-									<span className="text-red-500 text-xs">*</span>
-								</div>
-
-								<Field
-									id="did"
-									name="did"
-									className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
-									type="text"
-								/>
-								{formikHandlers?.errors?.did &&
-									formikHandlers?.touched?.did && (
-										<span className="text-red-500 absolute text-xs">
-											{formikHandlers?.errors?.did}
-										</span>
-									)}
-							</div>
-
-							<div className="mb-3 relative mt-3">
-								<label
-									htmlFor="method"
-									className="text-sm font-medium text-gray-900 dark:text-gray-300"
-								>
-									Method
-								</label>
-								<select
-									onChange={(e) => {
-										formikHandlers.handleChange(e);
-										handleLedgerChange(e);
-										setSelectedNetwork('');
-										setSelectedDid('');
-										setGeneratedKeys(null);
-									}}
-									id="method"
-									name="method"
-									className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500 h-11"
-								>
-									<option value="">Select Method</option>
-									{mappedData &&
-										Object.keys(mappedData)?.map((ledger) => (
-											<option key={ledger} value={ledger}>
-												{ledger}
-											</option>
-										))}
-								</select>
-								{formikHandlers?.errors?.method &&
-									formikHandlers?.touched?.method && (
-										<span className="absolute botton-0 text-red-500 text-xs">
-											{formikHandlers?.errors?.method}
-										</span>
-									)}
-							</div>
-
-							{formikHandlers.values.method === DidMethod.POLYGON && (
-			
-                                 <GenerateBtnPolygon  generatePolygonKeyValuePair={()=>generatePolygonKeyValuePair()}/>
-
-							)}
-
-							{generatedKeys && (
-								<div className="my-3 relative">
-									<p className="text-sm truncate">
-										<span className="font-semibold text-gray-900 dark:text-white">
-											Private Key:
-										</span>
-										<div className="flex ">
-											<CopyDid
-												className="align-center block text-sm text-gray-900 dark:text-white truncate"
-												value={generatedKeys?.privateKey}
-											/>
-										</div>
-									</p>
-
-									<p className="text-sm truncate">
-										<span className="font-semibold text-gray-900 dark:text-white">
-											Public Key Base58:
-										</span>
-										<div className="flex ">
-											<CopyDid
-												className="align-center block text-sm text-gray-900 dark:text-white truncate"
-												value={generatedKeys?.publicKeyBase58}
-											/>
-										</div>
-									</p>
-									<p className="text-sm truncate">
-										<span className="font-semibold text-gray-900 dark:text-white">
-											Address:
-										</span>
-										<div className="flex ">
-											<CopyDid
-												className="align-center block text-sm text-gray-900 dark:text-white truncate"
-												value={generatedKeys?.address}
-											/>
-										</div>
-									</p>
-								</div>
-							)}
-
-							{generatedKeys &&
-								formikHandlers.values.method === DidMethod.POLYGON && (
-
-									<TokenWarningMessage />
-								)}
-
-							{formikHandlers.values.method === DidMethod.POLYGON && (
-							
-								<SetPrivateKeyValue setPrivateKeyValue={(val:string)=>setPrivateKeyValue(val)} privateKeyValue={privateKeyValue}/>
-							)}
-
-							{formikHandlers.values.method === DidMethod.WEB && (
-
-								<SetDomainValueInput setDomainValue={(val:string)=>setDomainValue(val)} domainValue={domainValue}/>
-
-							)}
-
-							{formikHandlers.values.method !== DidMethod.POLYGON &&
-								formikHandlers.values.method !== DidMethod.KEY &&
-								formikHandlers.values.method !== DidMethod.WEB && (
-									<div className="my-3 relative">
-										<label
-											htmlFor="ledger"
-											className="text-sm font-medium text-gray-900 dark:text-gray-300"
-										>
-											Ledger
-										</label>
-										<select
-											onChange={(e) => {
-												formikHandlers.handleChange(e);
-												setSelectedNetwork(e.target.value);
-												setSelectedDid('');
-											}}
-											value={selectedNetwork}
-											id="ledger"
-											name="ledger"
-											className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500 h-11"
-										>
-											<option value="">Select Ledger</option>
-											{mappedData &&
-												selectedLedger &&
-												mappedData[selectedLedger] &&
-												Object.keys(mappedData[selectedLedger])?.map(
-													(ledger) => (
-														<option key={ledger} value={ledger}>
-															{ledger}
-														</option>
-													),
-												)}
-										</select>
-										{formikHandlers?.errors?.ledger &&
-											formikHandlers?.touched?.ledger && (
-												<span className="absolute botton-0 text-red-500 text-xs">
-													{formikHandlers?.errors?.ledger}
-												</span>
-											)}
-									</div>
-								)}
-							{formikHandlers.values.method !== DidMethod.WEB &&
-								formikHandlers.values.method !== DidMethod.KEY && (
-									
-									< LedgerLessMethodsComponents formikHandlers={formikHandlers} setSelectedDid={(val: string) => setSelectedDid(val)} selectedDid={selectedDid} mappedData={mappedData} selectedLedger={selectedLedger} selectedNetwork={selectedNetwork} />
-								)}
-
-							<div className="my-3 relative">
-								<label
-									htmlFor="DID Method"
-									className="text-sm font-medium text-gray-900 dark:text-gray-300"
-								>
-									DID Method
-								</label>
-								<div className="bg-gray-100 font-semibold text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500 h-11">
-									{showMethod(formikHandlers.values.method)}
-								</div>
-							</div>
-							<div className="mt-3 relative">
-								<Label htmlFor="name" value="Wallet Label" />
-								<span className="text-red-500 text-xs">*</span>
-
-								<Field
-									id="label"
-									name="label"
-									className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500"
-									type="text"
-								/>
-								{formikHandlers?.errors?.label &&
-									formikHandlers?.touched?.label && (
-										<span className="text-red-500 absolute text-xs">
-											{formikHandlers?.errors?.label}
-										</span>
-									)}
-							</div>
-							<div className="w-full flex justify-end">
-								<Button
-									type="submit"
-									className="flex h-min p-0.5 focus:z-10 focus:outline-none border border-transparent enabled:hover:bg-cyan-800 dark:enabled:hover:bg-cyan-700 mt-4 text-base font-medium text-center text-white bg-primary-700 rounded-md hover:!bg-primary-800 focus:ring-4 focus:ring-primary-300 sm:w-auto dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800"
-								>
-									Submit
-								</Button>
-							</div>
-						</Form>
-					)}
-				</Formik>
 			)}
 		</div>
 	);
@@ -1142,241 +549,6 @@ const WalletSpinup = (props: {
 					)}
 				</div>
 			</div>
-		</div>
-	);
-};
-
-const WalletSteps = (props: { steps: number; agentSpinupCall: boolean }) => {
-	return (
-		<div className="mt-4 ml-4">
-			<ol className="relative text-gray-500 border-l border-gray-200 dark:border-gray-700 dark:text-gray-400">
-				<li className="mb-10 ml-6">
-					{props.steps > 1 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-green-200 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-green-900">
-							<svg
-								className="w-3.5 h-3.5 text-green-500 dark:text-green-400"
-								aria-hidden="true"
-								xmlns="http://www.w3.org/2000/svg"
-								fill="none"
-								viewBox="0 0 16 12"
-							>
-								<path
-									stroke="currentColor"
-									strokeLinecap="round"
-									strokeLinejoin="round"
-									strokeWidth={2}
-									d="M1 5.917 5.724 10.5 15 1.5"
-								/>
-							</svg>
-						</span>
-					) : props.steps === 1 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700">
-							<svg
-								aria-hidden="true"
-								className="w-8 h-8 text-gray-200 animate-spin dark:text-gray-600 fill-blue-600"
-								viewBox="0 0 100 101"
-								fill="none"
-								xmlns="http://www.w3.org/2000/svg"
-							>
-								<path
-									d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z"
-									fill="currentColor"
-								/>
-								<path
-									d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z"
-									fill="currentFill"
-								/>
-							</svg>
-						</span>
-					) : (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-900 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700"></span>
-					)}
-					<h3 className="font-medium leading-tight">
-						Wallet creation is in progress
-					</h3>
-				</li>
-				<li className="mb-10 ml-6">
-					{props.steps > 2 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-green-200 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-green-900">
-							<svg
-								className="w-3.5 h-3.5 text-green-500 dark:text-green-400"
-								aria-hidden="true"
-								xmlns="http://www.w3.org/2000/svg"
-								fill="none"
-								viewBox="0 0 16 12"
-							>
-								<path
-									stroke="currentColor"
-									strokeLinecap="round"
-									strokeLinejoin="round"
-									strokeWidth={2}
-									d="M1 5.917 5.724 10.5 15 1.5"
-								/>
-							</svg>
-						</span>
-					) : props.steps === 2 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700">
-							<svg
-								aria-hidden="true"
-								className="w-8 h-8 text-gray-200 animate-spin dark:text-gray-600 fill-blue-600"
-								viewBox="0 0 100 101"
-								fill="none"
-								xmlns="http://www.w3.org/2000/svg"
-							>
-								<path
-									d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z"
-									fill="currentColor"
-								/>
-								<path
-									d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z"
-									fill="currentFill"
-								/>
-							</svg>
-						</span>
-						
-					) : (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700"></span>
-					)}
-					<h3 className="font-medium leading-tight">
-						Starting the DID publishing process... 
-					</h3>
-				</li>
-				<li className="mb-10 ml-6">
-					{props.steps > 3 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-green-200 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-green-900">
-							<svg
-								className="w-3.5 h-3.5 text-green-500 dark:text-green-400"
-								aria-hidden="true"
-								xmlns="http://www.w3.org/2000/svg"
-								fill="none"
-								viewBox="0 0 16 12"
-							>
-								<path
-									stroke="currentColor"
-									strokeLinecap="round"
-									strokeLinejoin="round"
-									strokeWidth={2}
-									d="M1 5.917 5.724 10.5 15 1.5"
-								/>
-							</svg>
-						</span>
-					) : props.steps === 3 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700">
-							<svg
-								aria-hidden="true"
-								className="w-8 h-8 text-gray-200 animate-spin dark:text-gray-600 fill-blue-600"
-								viewBox="0 0 100 101"
-								fill="none"
-								xmlns="http://www.w3.org/2000/svg"
-							>
-								<path
-									d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z"
-									fill="currentColor"
-								/>
-								<path
-									d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z"
-									fill="currentFill"
-								/>
-							</svg>
-						</span>
-					) : (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700"></span>
-					)}
-					<h3 className="font-medium leading-tight">
-						DID Publishing process is all done!
-					</h3>
-				</li>
-				<li className="mb-10 ml-6">
-					{props.steps > 4 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-green-200 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-green-900">
-							<svg
-								className="w-3.5 h-3.5 text-green-500 dark:text-green-400"
-								aria-hidden="true"
-								xmlns="http://www.w3.org/2000/svg"
-								fill="none"
-								viewBox="0 0 16 12"
-							>
-								<path
-									stroke="currentColor"
-									strokeLinecap="round"
-									strokeLinejoin="round"
-									strokeWidth={2}
-									d="M1 5.917 5.724 10.5 15 1.5"
-								/>
-							</svg>
-						</span>
-					) : props.steps === 4 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700">
-							<svg
-								aria-hidden="true"
-								className="w-8 h-8 text-gray-200 animate-spin dark:text-gray-600 fill-blue-600"
-								viewBox="0 0 100 101"
-								fill="none"
-								xmlns="http://www.w3.org/2000/svg"
-							>
-								<path
-									d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z"
-									fill="currentColor"
-								/>
-								<path
-									d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z"
-									fill="currentFill"
-								/>
-							</svg>
-						</span>
-					) : (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700"></span>
-					)}
-					<h3 className="font-medium leading-tight">
-						Creating an invitation URL...
-					</h3>
-				</li>
-				<li className="ml-6">
-					{props.steps > 5 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-green-200 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-green-900">
-							<svg
-								className="w-3.5 h-3.5 text-green-500 dark:text-green-400"
-								aria-hidden="true"
-								xmlns="http://www.w3.org/2000/svg"
-								fill="none"
-								viewBox="0 0 16 12"
-							>
-								<path
-									stroke="currentColor"
-									strokeLinecap="round"
-									strokeLinejoin="round"
-									strokeWidth={2}
-									d="M1 5.917 5.724 10.5 15 1.5"
-								/>
-							</svg>
-						</span>
-					) : props.steps === 5 ? (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700">
-							<svg
-								aria-hidden="true"
-								className="w-8 h-8 text-gray-200 animate-spin dark:text-gray-600 fill-blue-600"
-								viewBox="0 0 100 101"
-								fill="none"
-								xmlns="http://www.w3.org/2000/svg"
-							>
-								<path
-									d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z"
-									fill="currentColor"
-								/>
-								<path
-									d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z"
-									fill="currentFill"
-								/>
-							</svg>
-						</span>
-					) : (
-						<span className="absolute flex items-center justify-center w-8 h-8 bg-gray-100 rounded-full -left-4 ring-4 ring-white dark:ring-gray-900 dark:bg-gray-700"></span>
-					)}
-					<h3 className="font-medium leading-tight">
-						Invitation URL successfully created!
-					</h3>
-				</li>
-			</ol>
 		</div>
 	);
 };
